Improve unittests structure
1) Bugzilla server
We can not relay on status of bug from public bugzilla site.
Such bug can get closed, removed, or it may become unavailable.
I it better to feed cache with collection of our bugs.
Then we can be sure that bug has always same status as we expect.

2) 'test' target in setup.py
Added new target to setup.py to run tests.

python setup.py test

3) Test must give you information about quality
We can not accept result which has 2 pass and 2 failures and 1 skip,
and say that this is correct result of this test run.
Usual expected result of test is all test-cases passing.

Minor fix: Align setup.py with pep8